Create farmer.java
The farmer owns the farm and their main task is to choreograph all the activities on the farm. A good farmer is outstanding in their field, so they leave their house to stand in the middle of the field to be planted & harvested. They go home once the work is done.