Create coop.java
Collects the specified seed order from the seed bin, delivers it to the Coop Collection site and returns to the Coop Shed. The Coop will return once the Farmer requests a harvest pickup. The Coop will collect the harvest and deliver it to the harvest bin, then return to the Coop Shed.